feat(components) abstract components out of App.js
