+++ src/components/App.js
@@ -125,6 +125,9 @@ import AppControls from './AppControls'
 import AppControlBar from './AppControlBar'
 import AppHeader from './AppHeader'
 import AppContent from './AppContent'
+import AppSidebar from './AppSidebar'
+import AppPortfolio from './AppPortfolio'
+import AppPortfolioSettings from './AppPortfolioSettings'
 
 // Application Screens
 import LoadingScreen from './LoadingScreen'
@@ -717,396 +720,6 @@ class App extends React.Component {
     );
   }
 }
-
-
-class AppSidebar extends React.Component {
-  openPortfolioCreateScreen () {
-    CC.Events.emit('/screen/portfolio/create')
-  }
-
-  render () {
-    return (
-      <div className="layout-item sidebar">
-        <button onClick={this.openPortfolioCreateScreen}>Add Portfolio</button>
-        <AppSidebarPortfolioList
-          portfolios={this.props.portfolios} />
-      </div>
-    )
-  }
-}
-
-
-class AppSidebarPortfolioList extends React.Component {
-  render () {
-    return (
-      <ul className="portfolio-list">
-        { 
-          this.props.portfolios.map((portfolio, index) => {
-            return (
-              <li key={ portfolio.name }>
-                <AppSidebarPortfolioListItem
-                  index={index}
-                  portfolio={portfolio} />
-              </li>
-            )
-          })
-        }
-      </ul>
-    )
-  }
-}
-
-
-class AppSidebarPortfolioListItem extends React.Component {
-  componentDidMount() {
-    this.interval = setInterval(() => this.setState({ 
-      time: Date.now() 
-    }), 60000)
-  }
-
-  componentWillUnmount() {
-    clearInterval(this.interval)
-  }
-
-  openPortfolio (portfolioId) {
-    return (e) => {
-      CC.Events.emit('/screen/portfolio', {
-        portfolioId
-      })
-    }
-  }
-
-  render () {
-    if (this.props.portfolio.isOld) {
-      return (
-        <div className="portfolio-item not-draggable" onClick={ this.openPortfolio(this.props.index) }>
-          <div className="info">
-            <div className="title">
-              { this.props.portfolio.name }
-            </div>
-            <div className="last-updated">
-              This league is over, please remove.
-            </div>
-          </div>
-        </div>
-      )
-    }
-
-    let change = this.props.portfolio.getChange()
-    let holdings = this.props.portfolio.getHoldings()
-    let lastUpdated = this.props.portfolio.getLastUpdateTime()
-
-    return (
-      <div className="portfolio-item not-draggable" onClick={ this.openPortfolio(this.props.index) }>
-        <div className="info">
-          <div className="title">
-            { this.props.portfolio.name }
-          </div>
-          <div className="last-updated">
-            { lastUpdated }
-          </div>
-        </div>
-        
-        <div className="value">
-          <div className="total">
-            { holdings.valueFormatted } { holdings.currency }
-          </div>
-          <div className={`change ${ change.directionClassName }`}>
-           { change.directionIndicator } { change.valueFormatted } { change.currency }
-          </div>
-        </div>
-      </div>
-    )
-  }
-}
-
-class AppPortfolio extends React.Component {
-  componentDidMount() {
-    this.interval = setInterval(() => this.setState({ 
-      time: Date.now() 
-    }), 60000)
-  }
-
-  componentWillUnmount() {
-    clearInterval(this.interval)
-  }
-
-  render () {
-    return (
-      <div className="layout-content portfolio">
-        <AppPortfolioHeader
-          league={this.props.portfolio.league}
-          name={this.props.portfolio.name}
-          data={this.props.portfolio.history}
-          lastUpdated={this.props.portfolio.getLastUpdateTime()}
-          holdings={this.props.portfolio.getHoldings()}
-          change={this.props.portfolio.getChange()} />
-
-        <div className="items">
-          <table className="not-draggable">
-            <thead>
-              <tr>
-                <th><div>Icon</div></th>
-                <th><div>Item Name</div></th>
-                <th><div>Quantity</div></th>
-                <th><div>Value</div></th>
-              </tr>
-            </thead>
-            <tbody>
-            {this.props.portfolio.latestReport().items.map(entry => {
-              return entry.item.fullName ? (
-                <tr>
-                  <td>
-                    <img 
-                      src={entry.item.icon}
-                      width={32}
-                      style={{ verticalAlign: 'middle' }}
-                      title={entry.item.fullName}
-                    />
-                  </td>
-                  <td>{entry.item.fullName}</td>
-                  <td>{entry.stackSize}</td>
-                  <td>{entry.chaosValue.toFixed(2)} C</td>
-                </tr>
-              ) : null
-            })}
-            </tbody>
-          </table>
-        </div>
-      </div>
-    )
-  }
-}
-
-
-// Todo
-// - Add Chart Rendering after Load
-// - Remove Chart after unload
-class AppPortfolioChart extends React.Component {
-  render () {
-    return <div id="portfolio_chart" />
-  }
-}
-
-
-class AppPortfolioLargeStats extends React.Component {
-  render () {
-    return (
-      <div className="portfolio-large-stats">
-        <div>
-          <h3>{this.props.holdings.valueFormatted} {this.props.holdings.currency}</h3>
-          <h2>Holdings</h2>
-        </div>
-
-        <div className={`${this.props.profit.directionClassName}`}>
-          <h3>{this.props.profit.directionIndicator} {this.props.profit.valueFormatted} {this.props.profit.currency}</h3>
-          <h2>Last Gain / Loss</h2>
-        </div>
-
-        {/* <div>
-          <h3>{this.props.dayProfit} {this.props.currency}</h3>
-          <h2>24H Gain / Loss</h2>
-        </div> */}
-      </div>
-    )
-  }
-}
-
-
-class AppPortfolioMeta extends React.Component {
-  render () {
-    return (
-      <div className="portfolio-meta">
-        <div className="portfolio-meta-league">{ this.props.league }</div>
-        <div className="portfolio-meta-last-updated">{ this.props.lastUpdated }</div>
-      </div>
-    )
-  }
-}
-
-
-class AppPortfolioHeader extends React.Component {
-  render () {
-    return (
-      <div className="portfolio-header">
-        <AppPortfolioMeta
-          league={this.props.league}
-          lastUpdated={this.props.lastUpdated} />
-
-        <h1>{this.props.name}</h1>
-
-        <AppPortfolioLargeStats
-          holdings={this.props.holdings}
-          profit={this.props.change} />
-
-        <AppPortfolioChart
-          data={this.props.data} />
-      </div>
-    )
-  }
-}
-
-
-class AppPortfolioSettings extends React.Component {
-  state = {
-    settings: {
-      name: '',
-      league: '',
-      tracking: []
-    }
-  }
-
-  constructor (props) {
-    super(props)
-
-    this.handleChange = this.handleChange.bind(this)
-    this.handleLeagueChange = this.handleLeagueChange.bind(this)
-    this.handleTabChange = this.handleTabChange.bind(this)
-    this.handleSubmit = this.handleSubmit.bind(this)
-    this.generateName = this.generateName.bind(this)
-  }
-
-  generateName () {
-    let a = [ 'Strong', 'Fast', 'Powerful', 'Bold', 'Quick', 'Nimble', 
-              'Intense', 'Fixed', 'Long', 'Smart', 'Sweet', 'My', 'New' ]
-    let b = [ 'Exile', 'Tab', 'Folio', 'Tabs', 'Scion', 'Ranger', 'Witch', 
-              'Marauder', 'Shadow', 'Templar', 'Gangsta', 'Trade' ]
-
-    let choose = (list) => list[Math.floor(Math.random() * list.length)]
-    let name = `${choose(a)} ${choose(b)}`
-
-    this.setState({
-      settings: {
-        ...this.state.settings,
-        name
-      }
-    })
-  }
-
-  componentWillMount () {
-    if (!this.props.portfolio) {
-      this.generateName()
-      this.state.settings.league = this.props.leagues[0].id
-    }
-
-    this.setState({
-      settings: this.props.portfolio 
-        ? this.props.portfolio.settings 
-        : this.state.settings
-    })
-  }
-
-  handleChange (e) {
-    e.preventDefault()
-
-    const name = e.target.name
-    const value = e.target.type === 'checkbox' 
-      ? e.target.checked 
-      : e.target.value
-
-    this.setState({
-      settings: {
-        [name]: value
-      }
-    })
-  }
-
-  handleLeagueChange (option) {
-    let value = option ? option.value : ''
-
-    this.setState({
-      settings: {
-        ...this.state.settings,
-        tabs: null,
-        league: value
-      }
-    })
-  }
-
-  handleTabChange (tabs) {
-    this.setState({
-      settings: {
-        ...this.state.settings,
-        tabs
-      }
-    })
-  }
-
-  handleSubmit (e) {
-    e.preventDefault()
-
-    return this.props.portfolio
-      ? this.handleUpdate()
-      : this.handleCreate()
-  }
-
-  handleCreate () {
-    try {
-      CC.Events.emit('/portfolio/create', {
-        portfolio: this.state.settings
-      })
-      CC.Events.emit('/screen/dashboard')
-    } catch (e) {
-      console.log(e)
-    }
-  }
-
-  handleUpdate () {
-
-  }
-
-  render () {
-    let tabs = this.props.tabs[this.state.settings.league]
-
-    return (
-      <div className="portfolio-settings">
-        <h1>Portfolio Settings</h1>
-
-        <div className="form-group">
-          <h2>Portfolio Name</h2>
-          <input name="name" type="text" value={this.state.settings.name} onChange={this.handleChange} />
-          <Button onClick={this.generateName}>Generate?</Button>
-        </div>
-
-        <div className="form-group">
-          <h2>Portfolio League</h2>
-          <Select 
-            name="league"
-            clearable={false}
-            value={this.state.settings.league}
-            onChange={this.handleLeagueChange}
-            options={ this.props.leagues.map((league, index) => {
-              return {
-                value: league.id,
-                label: league.id
-              }
-            })} />
-        </div>
-
-        <div className="form-group">
-          <h2>Portfolio Tabs</h2>
-          <Select 
-            name="tabs"
-            multi={true}
-            value={this.state.settings.tabs}
-            onChange={this.handleTabChange}
-            options={tabs.map((tab, index) => {
-              return {
-                value: tab.id,
-                label: tab.name,
-                style: {
-                  color: `rgb(${tab.color[0]}, ${tab.color[1]}, ${tab.color[2]})`
-                }
-              }
-            })} />
-        </div>
-
-        <PrimaryButton onClick={this.handleSubmit}>Save</PrimaryButton>
-      </div>
-    )
-  }
-}
-
 
 // Capture uncaught errors
 process.on('uncaughtException', function (error) {
