+++ src/components/AppPortfolio/Header/Chart/index.js
@@ -1,0 +1,10 @@
+import React, { PropTypes } from 'react'
+import './index.css'
+
+class HeaderChart extends React.Component {
+  render () {
+    return <div id="portfolio_chart" />
+  }
+}
+
+export default HeaderChart
