+++ src/components/AppPortfolio/Header/index.css
@@ -1,0 +1,45 @@
+
+.portfolio-header {
+  background-image: linear-gradient(to bottom, #1e3c72 0%, #1e3c72 1%, #0D1821 100%);
+  color: #fff;
+  padding: 15px;
+  min-height: 200px;
+}
+
+.portfolio-header {
+  position: fixed;
+  width: 100%;
+  z-index: 1;
+}
+
+.portfolio-title {
+  display: flex;
+  justify-content: space-between;
+}
+
+.portfolio-header h3,
+.portfolio-header h2,
+.portfolio-header h1 {
+  font-weight: 400;
+  margin: 0;
+  padding: 0;
+}
+
+.portfolio-header h3 {
+  font-size: 24px;
+  font-family: 'Roboto Mono', monospace;
+}
+
+.portfolio-header h2 {
+  font-weight: 500;
+  color: hsla(210, 100%, 90%, 0.5);
+  font-size: 12px;
+  margin-top: 2px;
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+  line-height: 16px;
+}
+
+.portfolio-header #portfolio_chart {
+  margin: 0 -15px -15px;
+}
