+++ src/components/AppSidebarPortfolioList/index.css
@@ -1,0 +1,6 @@
+.portfolio-list,
+.portfolio-list li {
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
